refactor(recording): split replay and log-diff out of the recording manager
internal/recording held 16 files across three concerns: producing and storing
recordings, replaying them, and diffing two of them. The reference graph is a
clean layering — nothing in the manager, its store, or its quota accounting
calls into playback or logdiff; the traffic is entirely one-way — so the two
consumers move out:

  internal/recording          7  recording types, lifecycle, disk store, quotas
  internal/recording/playback 5  replay session, action execution, fragility
  internal/recording/logdiff  5  comparison, helpers, regression report

Note on evidence: every playback_engine_* file was created by one commit
(19ca6ede) and every logdiff_* file by another (4eecb893), so the filename
prefixes are provenance from those splits rather than independent evidence.
The call graph is the independent signal and agrees.

Where the evidence contradicted the filenames: playback.go held
IsFragileSelectorAction, which is a method on RecordingAction and belongs with
that type, not with the replay engine — it moved into types.go rather than into
the new playback package. Conversely logdiff_helpers.go held
CategorizeActionTypes as a *RecordingManager method that never touches its
receiver; it is now a plain function in logdiff.

Structural change: the two engines previously reached recording state through
methods on *RecordingManager. They now take a one-method interface each —
playback.RecordingSource (LookupRecording) and logdiff.RecordingSource
(GetRecording) — which *recording.Manager satisfies structurally. That removes
a delegation layer rather than adding one: internal/capture already was the
delegator, and its methods now call the engines directly.

LookupRecording is new and is the one method the replay path genuinely needed:
in-memory first, disk second, and deliberately without ValidateRecordingID,
because it addresses IDs the manager itself minted. StartPlayback used exactly
this lookup before; routing replay through GetRecording instead would have
started rejecting IDs that used to work.

Test moves and the coverage that came with them:
- logdiff and playback tests move with their code and now run against in-memory
  fakes instead of persisting to disk, so persistRecordingToDisk stays unexported.
- The manager's own load path (GetRecording) used to be covered only
  incidentally, via DiffRecordings in the log-diff tests. With those tests now
  on a fake it would have dropped to 0.0%, so manager_test.go gains direct
  coverage of GetRecording and LookupRecording. Verified RED under mutation
  (removing ValidateRecordingID from GetRecording; removing the in-memory
  branch from LookupRecording).
- playback gains TestExecuteRecordingDeletedAfterStart, which pins the
  playback_load_failed branch — the reason Execute re-reads the recording after
  Start rather than reusing it. That branch had no test before and would have
  silently disappeared in this refactor. Verified RED under mutation.
No function in either new subpackage reads 0.0%.

Proof of no behavior change: an end-to-end dump through internal/capture
(record -> stop -> get -> replay -> status -> fragility -> diff -> report ->
categorize, plus every error path) is byte-identical before and after.
