+++ docs/features/feature/playback-engine/tech-spec.md
@@ -10,9 +10,13 @@ last_verified_date: 2026-03-05
 # Playback Engine Tech Spec
 
 ## Architecture
-- Core types and orchestration: `internal/recording/playback.go`, `internal/recording/playback_engine_types.go`
-- Action execution helpers: `internal/recording/playback_engine_actions.go`
-- Session/runtime management: `internal/recording/playback_engine_session.go`
+- Recording data types and the manager that produces/persists them: `internal/recording/`
+- Replay session/result types and the `RecordingSource` seam: `internal/recording/playback/types.go`
+- Action execution helpers: `internal/recording/playback/actions.go`
+- Session/runtime management: `internal/recording/playback/session.go`
+- Selector fragility analysis: `internal/recording/playback/fragile.go`
+- Recording comparison and reports: `internal/recording/logdiff/`
+- Delegation surface: `internal/capture/recording_manager.go`
 - CLI handler bridge: `cmd/browser-agent/recording_handlers_playback.go`
 
 ## Constraints
