+++ internal/capture/recording_manager.go
@@ -1,5 +1,5 @@
-// Purpose: Re-exports recording-manager constructors and delegates recording lifecycle methods.
-// Why: Preserves capture package API compatibility while recording logic lives in internal/recording.
+// Purpose: Re-exports recording-manager constructors and delegates recording, replay, and log-diff calls.
+// Why: Preserves capture package API compatibility while recording logic lives in internal/recording{,/playback,/logdiff}.
 // Docs: docs/features/feature/playback-engine/index.md
 // Docs: docs/features/feature/tab-recording/index.md
 
@@ -7,6 +7,8 @@ package capture
 
 import (
 	"github.com/brennhill/Kaboom-Browser-AI-Devtools-MCP/internal/recording"
+	"github.com/brennhill/Kaboom-Browser-AI-Devtools-MCP/internal/recording/logdiff"
+	"github.com/brennhill/Kaboom-Browser-AI-Devtools-MCP/internal/recording/playback"
 )
 
 // NewRecordingManager creates a RecordingManager with initialized state.
@@ -42,34 +44,34 @@ func (c *Capture) GetRecording(recordingID string) (*Recording, error) {
 	return c.recordingManager.GetRecording(recordingID)
 }
 
-// StartPlayback delegates to RecordingManager.
+// StartPlayback delegates to the replay engine, reading recordings from RecordingManager.
 func (c *Capture) StartPlayback(recordingID string) (*PlaybackSession, error) {
-	return c.recordingManager.StartPlayback(recordingID)
+	return playback.Start(c.recordingManager, recordingID)
 }
 
-// ExecutePlayback delegates to RecordingManager.
+// ExecutePlayback delegates to the replay engine, reading recordings from RecordingManager.
 func (c *Capture) ExecutePlayback(recordingID string) (*PlaybackSession, error) {
-	return c.recordingManager.ExecutePlayback(recordingID)
+	return playback.Execute(c.recordingManager, recordingID)
 }
 
-// DetectFragileSelectors delegates to RecordingManager.
+// DetectFragileSelectors delegates to the replay engine.
 func (c *Capture) DetectFragileSelectors(sessions []*PlaybackSession) map[string]bool {
-	return c.recordingManager.DetectFragileSelectors(sessions)
+	return playback.DetectFragileSelectors(sessions)
 }
 
-// GetPlaybackStatus delegates to RecordingManager.
+// GetPlaybackStatus delegates to the replay engine.
 func (c *Capture) GetPlaybackStatus(session *PlaybackSession) map[string]any {
-	return c.recordingManager.GetPlaybackStatus(session)
+	return playback.Status(session)
 }
 
-// DiffRecordings delegates to RecordingManager.
+// DiffRecordings delegates to the log-diff engine, reading recordings from RecordingManager.
 func (c *Capture) DiffRecordings(originalID, replayID string) (*LogDiffResult, error) {
-	return c.recordingManager.DiffRecordings(originalID, replayID)
+	return logdiff.Compare(c.recordingManager, originalID, replayID)
 }
 
-// CategorizeActionTypes delegates to RecordingManager.
-func (c *Capture) CategorizeActionTypes(recording *Recording) map[string]int {
-	return c.recordingManager.CategorizeActionTypes(recording)
+// CategorizeActionTypes delegates to the log-diff engine.
+func (c *Capture) CategorizeActionTypes(rec *Recording) map[string]int {
+	return logdiff.CategorizeActionTypes(rec)
 }
 
 // GetStorageInfo delegates to RecordingManager.
