+++ internal/recording/logdiff/compare.go
@@ -1,38 +1,43 @@
 // Purpose: Compares original and replay recordings to produce a regression diff report.
 // Why: Separates high-level diff orchestration from helper utilities and report formatting.
-package recording
+package logdiff
 
-import "fmt"
+import (
+	"fmt"
 
-func (r *RecordingManager) DiffRecordings(originalRecordingID, replayRecordingID string) (*LogDiffResult, error) {
-	original, err := r.GetRecording(originalRecordingID)
+	"github.com/brennhill/Kaboom-Browser-AI-Devtools-MCP/internal/recording"
+)
+
+// Compare loads both recordings and diffs the replay against the original.
+func Compare(src RecordingSource, originalRecordingID, replayRecordingID string) (*Result, error) {
+	original, err := src.GetRecording(originalRecordingID)
 	if err != nil {
 		return nil, fmt.Errorf("logdiff_load_original_failed: Failed to load original recording: %w", err)
 	}
 
-	replay, err := r.GetRecording(replayRecordingID)
+	replay, err := src.GetRecording(replayRecordingID)
 	if err != nil {
 		return nil, fmt.Errorf("logdiff_load_replay_failed: Failed to load replay recording: %w", err)
 	}
 
-	result := &LogDiffResult{
+	result := &Result{
 		OriginalRecording: originalRecordingID,
 		ReplayRecording:   replayRecordingID,
-		NewErrors:         make([]DiffLogEntry, 0),
-		MissingEvents:     make([]DiffLogEntry, 0),
+		NewErrors:         make([]LogEntry, 0),
+		MissingEvents:     make([]LogEntry, 0),
 		ChangedValues:     make([]ValueChange, 0),
 	}
 
-	result.ActionStats = r.compareActions(original, replay)
-	r.detectRegressions(original, replay, result)
-	r.detectFixes(original, replay, result)
-	r.detectValueChanges(original, replay, result)
-	r.determineStatus(result)
+	result.ActionStats = compareActions(original, replay)
+	detectRegressions(original, replay, result)
+	detectFixes(original, replay, result)
+	detectValueChanges(original, replay, result)
+	determineStatus(result)
 
 	return result, nil
 }
 
-func (r *RecordingManager) compareActions(original, replay *Recording) ActionComparison {
+func compareActions(original, replay *recording.Item) ActionComparison {
 	stats := ActionComparison{
 		OriginalCount: original.ActionCount,
 		ReplayCount:   replay.ActionCount,
@@ -42,7 +47,7 @@ func (r *RecordingManager) compareActions(original, replay *Recording) ActionCom
 	return stats
 }
 
-func (r *RecordingManager) detectRegressions(original, replay *Recording, result *LogDiffResult) {
+func detectRegressions(original, replay *recording.Item, result *Result) {
 	originalErrors := make(map[string]bool)
 	for _, action := range original.Actions {
 		if action.Type == "error" {
@@ -52,7 +57,7 @@ func (r *RecordingManager) detectRegressions(original, replay *Recording, result
 
 	for _, action := range replay.Actions {
 		if action.Type == "error" && !originalErrors[action.Text] {
-			result.NewErrors = append(result.NewErrors, DiffLogEntry{
+			result.NewErrors = append(result.NewErrors, LogEntry{
 				Type:       "error",
 				Severity:   "high",
 				Level:      "error",
@@ -65,7 +70,7 @@ func (r *RecordingManager) detectRegressions(original, replay *Recording, result
 	}
 }
 
-func (r *RecordingManager) detectFixes(original, replay *Recording, result *LogDiffResult) {
+func detectFixes(original, replay *recording.Item, result *Result) {
 	replayErrors := make(map[string]bool)
 	for _, action := range replay.Actions {
 		if action.Type == "error" {
@@ -75,7 +80,7 @@ func (r *RecordingManager) detectFixes(original, replay *Recording, result *LogD
 
 	for _, action := range original.Actions {
 		if action.Type == "error" && !replayErrors[action.Text] {
-			result.MissingEvents = append(result.MissingEvents, DiffLogEntry{
+			result.MissingEvents = append(result.MissingEvents, LogEntry{
 				Type:       "error",
 				Severity:   "high",
 				Level:      "error",
@@ -88,7 +93,7 @@ func (r *RecordingManager) detectFixes(original, replay *Recording, result *LogD
 	}
 }
 
-func (r *RecordingManager) detectValueChanges(original, replay *Recording, result *LogDiffResult) {
+func detectValueChanges(original, replay *recording.Item, result *Result) {
 	originalValues := BuildTypeValueMap(original.Actions)
 
 	for _, action := range replay.Actions {
@@ -108,7 +113,7 @@ func (r *RecordingManager) detectValueChanges(original, replay *Recording, resul
 	}
 }
 
-func (r *RecordingManager) determineStatus(result *LogDiffResult) {
+func determineStatus(result *Result) {
 	if len(result.NewErrors) > 0 {
 		result.Status = "regression"
 		result.Summary = fmt.Sprintf("⚠️ REGRESSION: %d new errors detected", len(result.NewErrors))
